Add more tests of access rights in Layman and GS.
