+++ src/layman/common/prime_db_schema/publications_test.py
@@ -584,75 +584,3 @@ def test_publications_same_name():
     process_client.delete_map(username, publ_name)
     process_client.delete_layer(username2, publ_name)
     process_client.delete_map(username2, publ_name)
-
-
-@pytest.mark.usefixtures('ensure_layman')
-def test_rights_by_rest():
-    def test_by_type(prefix,
-                     publication_type,
-                     publish_method,
-                     patch_method,
-                     delete_method):
-        workspace_name = prefix + '_workspace'
-        username = prefix + '_user'
-        username2 = prefix + '_user2'
-        publication_name = prefix + '_publication'
-
-        with app.app_context():
-            workspaces.ensure_workspace(workspace_name)
-            id_workspace_user = workspaces.ensure_workspace(username)
-            users.ensure_user(id_workspace_user, userinfo)
-            id_workspace_user2 = workspaces.ensure_workspace(username2)
-            users.ensure_user(id_workspace_user2, userinfo)
-
-            publish_method(workspace_name,
-                           publication_name,
-                           access_rights={'read': f'{settings.RIGHTS_EVERYONE_ROLE}, {username2}',
-                                          'write': f'{settings.RIGHTS_EVERYONE_ROLE}, {username2}'})
-            info = publications.get_publication_infos(workspace_name, publication_type)[
-                (workspace_name, publication_type, publication_name)]
-            assert info['access_rights']['read'] == [username2, settings.RIGHTS_EVERYONE_ROLE]
-            assert info['access_rights']['write'] == [username2, settings.RIGHTS_EVERYONE_ROLE]
-
-            patch_method(workspace_name,
-                         publication_name,
-                         access_rights={'read': f'{settings.RIGHTS_EVERYONE_ROLE}'})
-            info = publications.get_publication_infos(workspace_name, publication_type)[
-                (workspace_name, publication_type, publication_name)]
-            assert info['access_rights']['read'] == [settings.RIGHTS_EVERYONE_ROLE]
-            assert info['access_rights']['write'] == [username2, settings.RIGHTS_EVERYONE_ROLE]
-
-            patch_method(workspace_name,
-                         publication_name,
-                         access_rights={'write': f'{settings.RIGHTS_EVERYONE_ROLE}'})
-            info = publications.get_publication_infos(workspace_name, publication_type)[
-                (workspace_name, publication_type, publication_name)]
-            assert info['access_rights']['read'] == [settings.RIGHTS_EVERYONE_ROLE]
-            assert info['access_rights']['write'] == [settings.RIGHTS_EVERYONE_ROLE]
-
-            patch_method(workspace_name,
-                         publication_name,
-                         access_rights={'read': f'{settings.RIGHTS_EVERYONE_ROLE}, {username}',
-                                        'write': f'{settings.RIGHTS_EVERYONE_ROLE}, {username}'})
-            info = publications.get_publication_infos(workspace_name, publication_type)[
-                (workspace_name, publication_type, publication_name)]
-            assert info['access_rights']['read'] == [username, settings.RIGHTS_EVERYONE_ROLE]
-            assert info['access_rights']['write'] == [username, settings.RIGHTS_EVERYONE_ROLE]
-
-            delete_method(workspace_name, publication_name)
-            users.delete_user(username)
-            users.delete_user(username2)
-            workspaces.delete_workspace(workspace_name)
-
-    test_by_type('test_rest_layer_rights',
-                 LAYER_TYPE,
-                 process_client.publish_layer,
-                 process_client.patch_layer,
-                 process_client.delete_layer,
-                 )
-    test_by_type('test_rest_map_rights',
-                 MAP_TYPE,
-                 process_client.publish_map,
-                 process_client.patch_map,
-                 process_client.delete_map,
-                 )
